Dom in separate module. Started with menu view
