+++ src/views/MenuView.js
@@ -1,0 +1,28 @@
+'use strict';
+
+import View from './View';
+
+/**
+ * Creates instance of MenuView
+ * 
+ * @class
+ * @classdesc Menu view. Render, shows, hide page.
+ */
+class MenuView extends View {
+	/**
+     * Creates instance of MenuView
+     */
+	constructor() {
+		super();
+	}
+    
+	show() {
+
+    }
+
+    hide() {
+        
+    }
+}
+
+export default MenuController;
