+++ src/views/View.js
@@ -1,149 +1,31 @@
 'use strict';
 
+import Dom from '../modules/Dom';
+
 /**
  * Creates instance of View
  * 
  * @class
- * @classdesc Base view. Provide work with DOM and creates base html template.
+ * @classdesc Base view. Creates base html template.
  */
 class View {
 	/**
-     * Creates instance of View
-     */
+	 * Creates instance of View
+	 */
 	constructor() {
 		if(View._instance) {
 			return View._instance;
 		}
 		View._instance = this;
-		this.prepareTemplate();
-	}
+		this._Dom = new Dom();
 
-	/**
-	 * Check if element has class.
-	 * 
-	 * @param {HTMLElement} element Element to check
-	 * @param {string} className Class name to check
-	 */
-	hasClass(element, className) {
-		return element.classList.contains(className);
-	}
+		this._body = this.Dom.getByTag(document, 'body')[0];
+		this._main = this.Dom.get(".block .main");
+		this._left = this.Dom.get(".block .left");
+		this._right = this.Dom.get(".block .right");
 
-	/**
-	 * Check if element has id.
-	 * 
-	 * @param {HTMLElement} element Element to check
-	 * @param {string} id Id to check
-	 */
-	hasId(element, id) {
-		return element.id === id;
-	}
+		this._parts = {};
 
-	/**
-	 * Check if element is tag.
-	 * 
-	 * @param {HTMLElement} element Element to check
-	 * @param {string} tagName Tag name to check
-	 */
-	isTag(element, tagName) {
-		element.tagName === tagName;
-	}
-
-	/**
-	 * Check if element has selector.
-	 * 
-	 * @param {HTMLElement} element Element to check
-	 * @param {string} selector Class name to check
-	 */
-	hasSelector(element, selector) {
-		const elementType = selector[0];
-		const s = selector.substring(1, selector.length);
-		switch(elementType) {
-		case '#':
-			return this.hasId(element, s);
-		case '.':
-			return this.hasClass(element, s);
-		default:
-			return this.isTag(element, selector);
-		}
-	}
-
-	/**
-	 * Search element by class name
-	 * 
-	 * @param {HTMLElement} element Root element
-	 * @param {string} className Class name to find
-	 */
-	getByClass(element, className) {
-		element.getElementsByClassName(className);
-	}
-
-	/**
-	 * Search element by id
-	 * 
-	 * @param {HTMLElement} element Root element
-	 * @param {string} id Id to find
-	 */
-	getById(element, id) {
-		element.getElementById(id);
-	}
-
-	/**
-	 * Search element by tag name
-	 * 
-	 * @param {HTMLElement} element Root element
-	 * @param {string} tagName Tag name to find
-	 */
-	getByTag(element, tagName) {
-		element.getElementsByTagName(tagName);
-	}
-
-	/**
-	 * Searching for element represented by selector.
-	 * 
-	 * @param {string} selector CSS selector like "tag #id .class"
-	 * @returns {(HtmlElement|Array|Boolean)} HTMLElement or array that contains elements. False if element not found.
-	 */
-	get(selector) {
-		let searchableElement = false;
-		const selectorArray = selector.split(' ');
-		selectorArray.forEach(selectorElement => {
-			const elementType = selectorElement[0];
-			const s = selectorElement.substring(1, selectorElement.length);
-			if(!searchableElement) {
-				switch(elementType) {
-				case '#':
-					searchableElement = this.getById(document, s);
-					break;
-				case '.':
-					searchableElement = this.getByClass(document, s);
-					break;
-				default:
-					searchableElement = this.getByTag(document, selectorElement);
-					break;
-				}
-				if(!searchableElement) {
-					return false;
-				}
-			} else {
-				if(searchableElement.isArray()) {
-					searchableElement = searchableElement.filter(element => {
-						return this.hasSelector(element, selectorElement);
-					});
-				} else {
-					if(!this.hasSelector(searchableElement, selectorElement)) {
-						return false;
-					}
-				}
-			}
-		});
-		return searchableElement;
-	}
-
-	/** 
-	 * Creates common html structure.
-	*/
-	prepareTemplate() {
-		document.getElementsByTagName('body')[0];
 	}
 
 	/** 
@@ -151,6 +33,29 @@ class View {
     */
 	show() {
 
+	}
+
+	/**
+	 * Hide page.
+	 */
+	hide() {
+		
+	}
+
+	get Dom() {
+		return this._Dom;
+	}
+
+	get main() {
+		return this._main;
+	}
+
+	get left() {
+		return this._left;
+	}
+
+	get right() {
+		return this._right;
 	}
 }
